+++ education_system/university_system/modules/shared/gui/main/main_gui.py
@@ -302,7 +302,6 @@ from education_system.university_system.modules.shared.gui.main.features.academi
     show_virtual_classroom_gui,
     show_office_hours_gui,
     show_ta_management_gui,
-    show_student_grades_gui,
     show_student_outcomes_gui,
     show_student_timetable_gui,
     show_student_registration_gui,
@@ -647,7 +646,6 @@ UnifiedManagementGUI.show_church_management_gui = show_church_management_gui
 UnifiedManagementGUI.show_bank_app_gui = show_bank_app_gui
 UnifiedManagementGUI.show_exam_scheduler_gui = show_exam_scheduler_gui
 UnifiedManagementGUI.show_student_analytics_gui = show_student_analytics_gui
-UnifiedManagementGUI.show_student_grades_gui = show_student_grades_gui
 UnifiedManagementGUI.show_student_outcomes_gui = show_student_outcomes_gui
 UnifiedManagementGUI.show_student_timetable_gui = show_student_timetable_gui
 UnifiedManagementGUI.show_student_registration_gui = show_student_registration_gui